pyproject.toml
[tool.poetry]
name = "harlequin"
version = "1.21.0"
description = "The SQL IDE for Your Terminal."
authors = ["Ted Conbeer <tconbeer@users.noreply.github.com>"]
license = "MIT"
homepage = "https://harlequin.sh"
repository = "https://github.com/tconbeer/harlequin"
readme = "README.md"
packages = [
    { include = "harlequin", from = "src" },
    { include = "harlequin_duckdb", from = "src" },
    { include = "harlequin_sqlite", from = "src" },
    { include = "harlequin_vscode", from = "src" },
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0.0"

# textual and component libraries
textual = "==0.56.4"
textual-fastdatatable = "==0.7.1"
textual-textarea = "==0.13.0"

# click
click = "^8.1.3"
rich-click = "^1.7.1"

# other deps
duckdb = ">=0.8.0"
shandy-sqlfmt = ">=0.19.0"
platformdirs = ">=3.10,<5.0"
importlib_metadata = { version = ">=4.6.0", python = "<3.10.0" }
tomli = { version = "^2.0.1", python = "<3.11.0" }
tomlkit = "^0.12.3"
questionary = "^2.0.1"

# temporary pins
# pin numpy to fix compatibility between pyarrow (textual-fastdatatable)
# and harlequin-databricks. Should be able to remove when harlequin-databricks
# dependency on databricks-sql-connector gets updated
numpy = [
    { version = "<1.25.0,>=1.21.0", python = ">=3.8.0,<3.9.0"},
    { version = "<2.0.0,>=1.21.0", python = ">=3.9.0,<3.13.0"}
]

# optional deps
boto3 = { version = "^1.34.22", optional = true }

# database adapters (optional installs for extras)
harlequin-postgres = { version = "^0.2", optional = true }
harlequin-mysql = { version = "^0.1", optional = true }
harlequin-odbc = { version = "^0.1", optional = true }
harlequin-bigquery = { version = "^1.0", optional = true }
harlequin-trino = { version = "^0.1", optional = true }
harlequin-databricks = { version = "^0.1", python = ">=3.9.0", optional = true }
harlequin-adbc = { version = "^0.1", python = ">=3.9.0", optional = true }
harlequin-cassandra = { version = "^0.1", python = ">=3.9.0", optional = true }
harlequin-nebulagraph = { version = "^0.1", python = ">=3.9.0", optional = true }

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.1"
textual-dev = "^1.0.1"
harlequin-postgres = "^0.2"
harlequin-mysql = "^0.1.1"
harlequin-odbc = "^0.1.1"
pyinstrument = "^4.6.2"

[tool.poetry.group.static.dependencies]
black = ">=23.3,<25.0"
ruff = ">=0.0.285"
mypy = "^1.2.0"
types-pygments = "^2.16.0.0"
pandas-stubs = "^2"
boto3-stubs = "^1.34.23"

[tool.poetry.group.test.dependencies]
pytest = ">=7.3.1,<9.0.0"
pytest-asyncio = "^0.21.0"
pytest-textual-snapshot = { git = "https://github.com/tconbeer/pytest-textual-snapshot.git", branch = "main" }
# extension tests require earlier version of duckdb.
duckdb = "^0.10"

[tool.poetry.scripts]
harlequin = "harlequin.cli:harlequin"

[tool.poetry.extras]
s3 = ["boto3"]
postgres = ["harlequin-postgres"]
mysql = ["harlequin-mysql"]
odbc = ["harlequin-odbc"]
bigquery = ["harlequin-bigquery"]
trino = ["harlequin-trino"]
databricks = ["harlequin-databricks"]
adbc = ["harlequin-adbc"]
cassandra = ["harlequin-cassandra"]
nebulagraph = ["harlequin-nebulagraph"]

[tool.poetry.plugins."harlequin.adapter"]
duckdb = "harlequin_duckdb:DuckDbAdapter"
sqlite = "harlequin_sqlite:HarlequinSqliteAdapter"

[tool.poetry.plugins."harlequin.keymap"]
vscode = "harlequin_vscode:VSCODE"

[tool.poetry.plugins."pygments.styles"]
harlequin = "harlequin.colors:HarlequinPygmentsStyle"

[tool.ruff]
target-version = "py38"

[tool.ruff.lint]
select = ["A", "B", "E", "F", "I"]

[tool.mypy]
python_version = "3.8"
files = [
    "src/**/*.py",
    "tests/**/*.py",
]
mypy_path = "src:stubs"

show_column_numbers = true

# show error messages from unrelated files
follow_imports = "normal"

# be strict
disallow_untyped_calls = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
strict_optional = true

warn_return_any = true
warn_no_return = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true

no_implicit_reexport = true
strict_equality = true

[tool.pytest.ini_options]
markers = [
    "online: requeries internet connection (deselect with '-m \"not online\"')",
    "use_cache: override the autouse fixture the disables the buffer cache",
    "flaky_windows: likely to fail for no reason on windows."
]
